backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
mysqlclient>=2.2
python-decouple>=3.8
gunicorn>=21.2
